Initial import of cleanup work
