+++ QUICKSTART.md
@@ -1,4 +1,5 @@
 # Fabric Quickstart
+For production deploys and native dependencies, see [`docs/PRODUCTION.md`](docs/PRODUCTION.md).  Contributor-oriented detail is in [`DEVELOPERS.md`](DEVELOPERS.md).
 
 ## Prerequisites
 0. (optional) Install NVM: `https://nvm.sh`
@@ -63,10 +64,10 @@ Finally, run the program:
 ## Setup Local Bitcoin Node
 For testing purposes, we recommend running `bitcoin-qt` using the bash command with `scripts/bitcoin-playnet.sh`. Follow these steps to get setup:
 
-1. Download the [bitcoin-core client](https://bitcoin.org/bin/) to your system. 
+1. Download the [bitcoin-core client](https://bitcoin.org/bin/) to your system.
 2. (Recommended) Download SHA256SUMS.sig and SHA256SUMS to verify hashes match, using these commands in the terminal:
 	```
-	$ cat SHA256SUMS # view the sha-256 file hashes 
+	$ cat SHA256SUMS # view the sha-256 file hashes
 	$ sha256sum downloaded_file.name # compute hash of downloaded file
 	```
 3. (Recommended) Verify hashes with a Bitcoin Core Release Signing Key, keys to respective versions can be found [here](https://bitcoin.org/en/download).
@@ -74,20 +75,20 @@ For testing purposes, we recommend running `bitcoin-qt` using the bash command w
 	$ gpg --verify SHA256SUMS.sig SHA256SUMS # verify file is signed
 	```
 4. Continue with the bitcoin-core installation to your system.
-5. Check installation in terminal by running `bitcoind --version` or `bitcoin-qt --version`. If the terminal outputs something like `Bitcoin Core version vXX.X.X` then you are ready to run the scripts 
-6. Navigate your terminal window to the [fabric](https://github.com/fabriclabs/fabric) directory and create a new directory named "bitcoin-playnet" inside "stores" 
+5. Check installation in terminal by running `bitcoind --version` or `bitcoin-qt --version`. If the terminal outputs something like `Bitcoin Core version vXX.X.X` then you are ready to run the scripts
+6. Navigate your terminal window to the [fabric](https://github.com/fabriclabs/fabric) directory and create a new directory named "bitcoin-playnet" inside "stores"
 	```
 	$ cd stores && mkdir bitcoin-playnet
 	```
-	There should now be a directory path `/fabric/stores/bitcoin-playnet`. 	
+	There should now be a directory path `/fabric/stores/bitcoin-playnet`.
 	Return to the parent directory of the repo `cd ..`, now you are ready to run the script
-7. Run the playnet bash script with 
+7. Run the playnet bash script with
 	```
 	$ ./scripts/bitcoin-playnet.sh
 	```
 	A bitcoin client should now pop up
 8. Create or restore a dev wallet (not to be used with real funds)
-9. (optional) view the debug log with this command in a new terminal window 
+9. (optional) view the debug log with this command in a new terminal window
 	```
 	$ tail -f stores/bitcoin-playnet/regtest/debug.log
 	```
