+++ docs/README.md
@@ -1,2 +1,13 @@
-# Basic Docs
-foo
+# Documentation index — `@fabric/core`
+
+| Doc | Audience | Purpose |
+|-----|------------|---------|
+| [../README.md](../README.md) | Everyone | Install, CLI, API snippet, ecosystem |
+| [../DEVELOPERS.md](../DEVELOPERS.md) | Contributors | Repo layout, tests, architecture, storage |
+| [../QUICKSTART.md](../QUICKSTART.md) | New users | Minimal install → `fabric chat` |
+| [../AGENTS.md](../AGENTS.md) | Service authors | Agent lifecycle, workers, safety |
+| [PRODUCTION.md](PRODUCTION.md) | Operators | Node version, native deps, release gate |
+| [../PRIVACY.md](../PRIVACY.md) | Operators | Threat model |
+| [../SECURITY.md](../SECURITY.md) | Reporters | Process, disclosure |
+
+Generated after `npm run make:docs`: HTML under `docs/*.html` (JSDoc) plus root `API.md` from `npm run make:api`.
