+++ types/entity.js
@@ -2,6 +2,8 @@
 
 const crypto = require('crypto');
 const { EventEmitter } = require('events');
+const monitor = require('fast-json-patch');
+const Witness = require('./witness');
 
 class Entity extends EventEmitter {
   constructor (data = {}) {
@@ -173,4 +175,94 @@ class Entity extends EventEmitter {
   }
 }
 
+class Transition extends Entity {
+  constructor (settings = {}) {
+    super(settings);
+
+    this.status = 'constructing';
+    this._state = {
+      origin: null,
+      target: null,
+      changes: [],
+      program: [],
+      witness: null
+    };
+
+    this.settings = Object.assign({}, this._state, settings);
+    this.witness = new Witness(this.settings);
+
+    this._setOrigin(this.settings.origin);
+    this._setTarget(this.settings.target);
+
+    this.status = 'constructed';
+  }
+
+  static between (origin, target) {
+    const x = new Entity(origin);
+    const y = new Entity(target);
+
+    const actor = Object.assign({}, origin);
+    const observer = monitor.observe(actor);
+
+    Object.assign(actor, target);
+
+    return new Transition({
+      origin: x.id,
+      target: y.id,
+      changes: monitor.generate(observer)
+    });
+  }
+
+  fromTarget (target) {
+    const base = new Entity();
+    const entity = this._describeTarget(target);
+    return new Transition({
+      origin: base.id,
+      target: entity.id
+    });
+  }
+
+  _applyTo (state) {
+    if (!state) throw new Error('State must be provided.');
+    if (!(state instanceof Entity)) throw new Error('State not of known Entity type.');
+
+    const instance = Object.assign({}, state);
+    const observer = monitor.observe(instance);
+
+    try {
+      monitor.applyPatch(instance, this._state.changes);
+    } catch (E) {
+      console.error('Could not apply changes:', E);
+    }
+
+    monitor.generate(observer);
+    return instance;
+  }
+
+  _setChanges (changes) {
+    if (!changes) throw new Error('No changes specified.');
+    this._state.changes = changes;
+  }
+
+  _setOrigin (origin) {
+    if (!origin) throw new Error('No origin specified.');
+    this._state.origin = origin;
+  }
+
+  _setTarget (target) {
+    if (!target) throw new Error('No target specified.');
+    this._state.target = target;
+  }
+
+  _describeTarget (target) {
+    if (!target) throw new Error('No target specified.');
+    const entity = new Entity(target);
+    this._setTarget(entity.id);
+    return entity;
+  }
+}
+
+Entity.Transition = Transition;
+
 module.exports = Entity;
+module.exports.Transition = Transition;
