+++ types/fabric.d.ts
@@ -1,0 +1,7 @@
+/**
+ * Minimal entry typings for `@fabric/core`. The runtime is CommonJS (`types/fabric.js`);
+ * leaf modules under `types/*.js` are authoritative. Expand declarations as coverage grows
+ * (see REMAINING_WORK.md).
+ */
+declare const Fabric: any;
+export = Fabric;
