--- types/keystore.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const Store = require('./store');
-
-class Keystore extends Store {
-  constructor (settings = {}) {
-    super(Store.encryptedSettings(settings));
-  }
-}
-
-module.exports = Keystore;
